refactor(scene3d): 抽出物体/人偶/人群渲染图元到 scene3dObjects.tsx [split 3/N]
Scene3DMeshGeometry/Mannequin/Crowd/FootRings 等渲染图元+几何 helper 移出;
壳 2864→2246。双 tsc 零错误
